backend/data_fetcher/requirements.txt
pymongo
requests
python-dotenv
schedule
